requirements.txt
fpdf2==2.7.8
Pygments==2.18.0
